requirements.txt
streamlit==1.26.0
requests==2.31.0
beautifulsoup4==4.13.5
lxml==4.9.3
PyMuPDF==1.22.5
python-docx==0.8.11
langchain==0.3.27
openai==1.33.0
